% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{rosy}
\alias{rosy}
\title{Authenticate with the Roswell Park Data Commons}
\usage{
rosy(user, credential)
}
\arguments{
\item{user}{character(1) user identity}

\item{credential}{character(1) path to file containing
autehtnication credentials}
}
\value{
An object of class `Rosy` containing credentials and
    current state of the user session.
}
\description{
Authenticate with the Roswell Park Data Commons
}
