% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSupers.R
\name{call_super_cutoff_internal}
\alias{call_super_cutoff_internal}
\title{Call 'supers' on a vector of values.}
\usage{
call_super_cutoff_internal(vec, target_slope = (max(vec) -
  min(vec))/length(vec))
}
\description{
Call 'supers' on a vector of values.
}
