% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallSupers.R
\name{call_supers_internal}
\alias{call_supers_internal}
\title{Call 'supers' on a RSE. Adds to an assay slot named 'isSuper.'}
\usage{
call_supers_internal(rse, set_rnk = F, assay = "readInPeakNormPerBP")
}
\description{
Call 'supers' on a RSE. Adds to an assay slot named 'isSuper.'
}
