% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equivalentAssays}
\alias{equivalentAssays}
\title{Check two RangedSummarizedExperiments have the same
assays.}
\usage{
equivalentAssays(rse1, rse2)
}
\arguments{
\item{rse1}{A RangedSummarizedExperiment}

\item{rse2}{A RangedSummarizedExperiment}
}
\description{
Check two RangedSummarizedExperiments have the same
assays.
}
