% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equivalentRanges}
\alias{equivalentRanges}
\title{Check two RangedSummarizedExperiments use the same ranges}
\usage{
equivalentRanges(rse1, rse2)
}
\arguments{
\item{rse1}{A RangedSummarizedExperiment}

\item{rse2}{A RangedSummarizedExperiment}
}
\description{
Check two RangedSummarizedExperiments use the same ranges
}
