% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stitching.R
\name{extend_merge_internal}
\alias{extend_merge_internal}
\title{Extend GRanges by the specified distance}
\usage{
extend_merge_internal(gr, width = 0)
}
\arguments{
\item{gr}{A GRanges object.}

\item{sd}{Stitching distance.}
}
\description{
A wrapper around `GenomicRanges::flank`.
}
\details{
Returns a GRanges object with flanking regions added to every distinct
region. Metadata and strand info is lost in this process.
}
