% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stitching.R
\name{get_grl_entry_width_internal}
\alias{get_grl_entry_width_internal}
\title{Get the total with of a GRL element.}
\usage{
get_grl_entry_width_internal(x)
}
\description{
Get the total with of a GRL element.
}
