% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stitching.R
\name{get_stitch_stats_internal}
\alias{get_stitch_stats_internal}
\title{Get the stitch stats for a range of distances so rosy can
internally choose an optimal distance.}
\usage{
get_stitch_stats_internal(gr)
}
\description{
Get the stitch stats for a range of distances so rosy can
internally choose an optimal distance.
}
