% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChrNameFromGr}
\alias{getChrNameFromGr}
\title{Get 1 chr name from a GR and confirm all are from same level.}
\usage{
getChrNameFromGr(gr)
}
\description{
Get 1 chr name from a GR and confirm all are from same level.
}
