% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liquidating.R
\name{normalize_internal}
\alias{normalize_internal}
\title{Scale signal in RangedSummarizedExperiment.}
\usage{
normalize_internal(rse)
}
\arguments{
\item{rse}{A RangedSummarizedExperiment}
}
\value{
RangedSummarizedExperiment
}
\description{
Scale signal in RangedSummarizedExperiment.
}
