% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liquidating.R
\name{normalize_mrip_internal}
\alias{normalize_mrip_internal}
\title{Scale signal in RangedSummarizedExperiment by MRIP.}
\usage{
normalize_mrip_internal(rse, assay = "counts")
}
\arguments{
\item{rse}{A RangedSummarizedExperiment}
}
\value{
RangedSummarizedExperiment
}
\description{
Scale signal in RangedSummarizedExperiment by MRIP.
}
