% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Liquidating.R
\name{normalize_perbp_internal}
\alias{normalize_perbp_internal}
\title{Scale signal in RangedSummarizedExperiment per BP.}
\usage{
normalize_perbp_internal(rse, assay = "readInPeakNorm")
}
\arguments{
\item{rse}{A RangedSummarizedExperiment}
}
\value{
RangedSummarizedExperiment
}
\description{
Scale signal in RangedSummarizedExperiment per BP.
}
