% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{filter.Rosy}
\alias{filter.Rosy}
\alias{select.Rosy}
\title{Discover resources available in the Roswell Park Data Commons}
\usage{
\method{filter}{Rosy}(.data, ...)

\method{select}{Rosy}(.data, ...)
}
\arguments{
\item{.data}{An object of class `Rosy`, created with `rosy()`.}

\item{...}{Additional arguments determining data set filters and
date type selections.}
}
\value{
An updated object of class `Rosy`, restricting available
    data to specific data sets (`filter()`) and columsn
    (`select()`).
}
\description{
Discover resources available in the Roswell Park Data Commons
}
