% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{as_tibble.Rosy}
\alias{as_tibble.Rosy}
\title{Retrieve data from the Roswell Park Data Commons}
\usage{
\method{as_tibble}{Rosy}(x, ...)
}
\arguments{
\item{x}{An object of class `Rosy`, created and authenticated with
`rosy()` with subsets defined via `filter()` and `select()`.}

\item{...}{additional arguments; ignored.}
}
\value{
A tibble containing desired data.
}
\description{
Retrieve data from the Roswell Park Data Commons
}
