% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosy.R
\name{rose_single_internal}
\alias{rose_single_internal}
\title{Run rose stitching + quant for a single sample.}
\usage{
rose_single_internal(regions, reads, paired = F, colData = NULL,
  stitchDist = 12500, se_cutoff = "optimize")
}
\arguments{
\item{regions}{GRanges or path to bed.}

\item{reads}{Object or path to file containing reads.}

\item{paired}{Logical indicating paired-end reads.}

\item{colData}{dataframe with sample information.}
}
\description{
Will take only a single set of regions and reads.
}
