% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosy-class.R
\docType{class}
\name{Rosy-class}
\alias{Rosy-class}
\alias{.Rosy}
\alias{show,Rosy-method}
\title{Class representation of a Roswll Park Data Commons session}
\usage{
\S4method{show}{Rosy}(object)
}
\arguments{
\item{object}{An object of class `Rosy`.}
}
\value{
`show()` is invoked implicitly for the side-effect of
    displaying the object.
}
\description{
Class representation of a Roswll Park Data Commons session
}
