% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{rosy}
\alias{rosy}
\alias{rosy,GRanges,character,ANY,data.frame-method}
\title{Run rosy.}
\usage{
rosy(regions, reads, paired = F, colData, txdb = NULL,
  stitchDist = 12500, se_cutoff = "optimize", exclude = NULL)

\S4method{rosy}{GRanges,character,ANY,data.frame}(regions, reads,
  paired = F, colData, txdb = NULL, stitchDist = 12500,
  se_cutoff = "optimize", exclude = NULL)
}
\arguments{
\item{regions}{GRanges, list of GRanges, GRangesList, path to file, or list of paths.}

\item{reads}{Object or path to file containing reads, or list.}

\item{paired}{Logical indicating paired-end reads or list.}

\item{colData}{dataframe with sample information.}

\item{stitchDist}{Enhancer stitching distance.}

\item{se_cutoff}{Either 'optimize' or an integer indicating the top n enhancers to consider as SE.}

\item{exclude}{GRanges of regions to exclude}
}
\description{
Run rosy.
}
