% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosy.R
\name{rosy_internal}
\alias{rosy_internal}
\title{Run rosy for a single sample.}
\usage{
rosy_internal(regions, reads, paired = F, colData = NULL,
  stitchDist = 12500)
}
\arguments{
\item{regions}{GRanges.}

\item{reads}{Object or path to file containing reads.}

\item{paired}{Logical indicating paired-end reads.}

\item{colData}{dataframe with sample information.}
}
\description{
Will take only a single set of regions and reads.
}
