% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{sds2tbl}
\alias{sds2tbl}
\title{rdname sds2tbl}
\usage{
sds2tbl(sds, txdb = "TxDb.Hsapiens.UCSC.hg19.knownGene",
  bsgenome = NULL)
}
\description{
Convert an SE dataset to a tibble for plotting and downstream
querying.
}
