% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate.R
\name{sds2tbl_single_internal}
\alias{sds2tbl_single_internal}
\title{Convert an SE dataset to a tibble for plotting and downstream
querying.}
\usage{
sds2tbl_single_internal(sds)
}
\description{
Convert an SE dataset to a tibble for plotting and downstream
querying.
}
