% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Rosy.R
\docType{methods}
\name{stitch}
\alias{stitch}
\alias{stitch,GRanges,numeric-method}
\alias{stitch,GRangesList,numeric-method}
\alias{stitch,character,numeric-method}
\alias{rosy}
\title{Stitch ranges within `stitchDist` of each other.}
\usage{
stitch(object, stitchDist = 12500)

\S4method{stitch}{GRanges,numeric}(object, stitchDist = 12500)

\S4method{stitch}{GRangesList,numeric}(object, stitchDist = 12500)

\S4method{stitch}{character,numeric}(object, stitchDist = 12500)

rosy(regions, reads, paired = F, colData = None)
}
\arguments{
\item{object}{GRanges, GRangesList, or paths to files.}

\item{regions}{GRanges, GRangesList, or paths to files.}

\item{reads}{Objects or paths to file containing reads.}

\item{paired}{Logical indicating paired-end reads.}

\item{colData}{dataframe with sample information.}
}
\description{
Stitch ranges within `stitchDist` of each other.

Run rosy.
}
