% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stitching.R
\name{stitch_internal}
\alias{stitch_internal}
\title{Stitch GRanges within specified distance}
\usage{
stitch_internal(gr, sd = 0)
}
\arguments{
\item{gr}{A GRanges object.}

\item{sd}{Stitching distance.}
}
\description{
Returns a GRanges object stitched together such that
ranges within the specified stitch distance of each other
are now reduced. In the process, metadata  and
strand information is lost.
}
