% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindRseCols}
\alias{bindRseCols}
\title{Combine RangedSummarizedExperiments}
\usage{
bindRseCols(rse1, rse2)
}
\arguments{
\item{rse1}{RangedSummarizedExperiment}

\item{rse2}{RangedSummarizedExperiment}
}
\description{
Combine RangedSummarizedExperiments
}
