% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{rosy}
\alias{rosy}
\alias{rosy,character,character,logical,data.frame,TxDb,numeric-method}
\title{Run rosy.}
\usage{
rosy(regions, reads, paired, colData, txdb, stitchDist = 12500)


  \S4method{rosy}{character,character,logical,data.frame,TxDb,numeric}(regions,
  reads, paired, colData, txdb, stitchDist = 12500)
}
\arguments{
\item{regions}{GRanges, GRangesList, or path to file.}

\item{reads}{Object or path to file containing reads.}

\item{paired}{Logical indicating paired-end reads.}

\item{colData}{dataframe with sample information.}
}
\description{
Run rosy.
}
