% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosy.R
\name{rosy_internal}
\alias{rosy_internal}
\title{Run rosy on a single sample.}
\usage{
rosy_internal(regions, reads, paired = F, colData = NULL, txdb = NULL,
  stitchDist = "optimize")
}
\arguments{
\item{regions}{GRanges.}

\item{reads}{Object or path to file containing reads.}

\item{paired}{Logical indicating paired-end reads.}

\item{colData}{dataframe with sample information.}
}
\description{
Run rosy on a single sample.
}
