% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{stitch}
\alias{stitch}
\alias{stitch,GRanges,numeric-method}
\alias{stitch,GRangesList,numeric-method}
\alias{stitch,character,numeric-method}
\title{Stitch ranges within `stitchDist` of each other.}
\usage{
stitch(object, stitchDist = 12500)

\S4method{stitch}{GRanges,numeric}(object, stitchDist = 12500)

\S4method{stitch}{GRangesList,numeric}(object, stitchDist = 12500)

\S4method{stitch}{character,numeric}(object, stitchDist = 12500)
}
\arguments{
\item{object}{GRanges, GRangesList, or paths to files.}
}
\description{
Stitch ranges within `stitchDist` of each other.
}
