% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{stitch}
\alias{stitch}
\alias{stitch,GRanges,numeric-method}
\title{Stitch ranges within stitchDist.}
\usage{
stitch(object, stitchDist = 12500)

\S4method{stitch}{GRanges,numeric}(object, stitchDist = 12500)
}
\arguments{
\item{object}{GRanges object}
}
\description{
Stitch ranges within stitchDist.
}
